CVM-7: Limit permission for companies

+++ src/Controller/ResumeController.php
@@ -5,17 +5,24 @@ namespace App\Controller;
 use App\Entity\VacancyResume;
 use App\Form\ResumeType;
 use App\Repository\VacancyResumeRepository;
+use Doctrine\Common\Collections\Collection;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 #[Route('/resume', name: 'resume_')]
 class ResumeController extends AbstractController
 {
+    private ?Collection $allowedVacancies = null;
+
     public function __construct(
+        private readonly TokenStorageInterface $tokenStorage,
         private readonly VacancyResumeRepository $resumeRepository,
-    ) {}
+    ) {
+        $this->checkAvailableVacancies();
+    }
 
     #[Route('/new', name: 'new')]
     public function new(
@@ -23,7 +30,7 @@ class ResumeController extends AbstractController
     ): Response
     {
         $company = new VacancyResume();
-        $form = $this->createForm(ResumeType::class, $company);
+        $form = $this->createForm(ResumeType::class, $company, ['allowed_vacancies' => $this->allowedVacancies]);
 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -56,7 +63,11 @@ class ResumeController extends AbstractController
         VacancyResume $resume,
     ): Response
     {
-        $form = $this->createForm(ResumeType::class, $resume);
+        if (!$this->allowedVacancies->contains($resume->getVacancy())) {
+            return $this->redirectToRoute('resume_list');
+        }
+
+        $form = $this->createForm(ResumeType::class, $resume, ['allowed_vacancies' => $this->allowedVacancies]);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $resume->updateDate();
@@ -80,8 +91,15 @@ class ResumeController extends AbstractController
     )]
     public function delete(VacancyResume $resume): Response
     {
-        $this->resumeRepository->remove($resume, true);
+        if ($this->allowedVacancies->contains($resume->getVacancy())) {
+            $this->resumeRepository->remove($resume, true);
+        }
 
         return $this->redirectToRoute('resume_list');
+    }
+
+    private function checkAvailableVacancies(): void
+    {
+        $this->allowedVacancies = $this->tokenStorage->getToken()->getUser()->getCompany()?->getVacancies();
     }
 }
